flytta på funktionen calculateMiddleColor
